third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-element-offsets-expected.txt

This is a testharness.js-based test.
Found 54 tests; 7 PASS, 47 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL Scroll at offset 0 updates animation correctly [element_start_start] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 199 updates animation correctly [element_start_start] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 200 updates animation correctly [element_start_start]
FAIL Scroll at offset 300 updates animation correctly [element_start_start] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 398 updates animation correctly [element_start_start] assert_equals: expected "199px" but got "100px"
FAIL Scroll at offset 400 updates animation correctly [element_start_start] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_end_end] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 49 updates animation correctly [element_end_end] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 50 updates animation correctly [element_end_end]
FAIL Scroll at offset 150 updates animation correctly [element_end_end] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 248 updates animation correctly [element_end_end] assert_equals: expected "199px" but got "100px"
FAIL Scroll at offset 250 updates animation correctly [element_end_end] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_end_start] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 49 updates animation correctly [element_end_start] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 50 updates animation correctly [element_end_start]
FAIL Scroll at offset 225 updates animation correctly [element_end_start] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 393 updates animation correctly [element_end_start] assert_equals: expected "198px" but got "100px"
FAIL Scroll at offset 400 updates animation correctly [element_end_start] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_end_1_end] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 99 updates animation correctly [element_end_1_end] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 100 updates animation correctly [element_end_1_end]
FAIL Scroll at offset 175 updates animation correctly [element_end_1_end] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 247 updates animation correctly [element_end_1_end] assert_equals: expected "198px" but got "100px"
FAIL Scroll at offset 250 updates animation correctly [element_end_1_end] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_end_start_05] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 49 updates animation correctly [element_end_start_05] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 50 updates animation correctly [element_end_start_05]
FAIL Scroll at offset 206 updates animation correctly [element_end_start_05] assert_equals: expected "148px" but got "100px"
FAIL Scroll at offset 362 updates animation correctly [element_end_start_05] assert_equals: expected "196px" but got "100px"
FAIL Scroll at offset 375 updates animation correctly [element_end_start_05] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_start_400px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 199 updates animation correctly [element_start_400px] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 200 updates animation correctly [element_start_400px]
FAIL Scroll at offset 300 updates animation correctly [element_start_400px] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 398 updates animation correctly [element_start_400px] assert_equals: expected "199px" but got "100px"
FAIL Scroll at offset 400 updates animation correctly [element_start_400px] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_50px_end] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 49 updates animation correctly [element_50px_end] assert_equals: expected "0px" but got "100px"
PASS Scroll at offset 50 updates animation correctly [element_50px_end]
FAIL Scroll at offset 150 updates animation correctly [element_50px_end] assert_equals: expected "150px" but got "100px"
FAIL Scroll at offset 248 updates animation correctly [element_50px_end] assert_equals: expected "199px" but got "100px"
FAIL Scroll at offset 250 updates animation correctly [element_50px_end] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_outside] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 100 updates animation correctly [element_outside] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 200 updates animation correctly [element_outside] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 300 updates animation correctly [element_outside] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 400 updates animation correctly [element_outside] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 450 updates animation correctly [element_outside] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 0 updates animation correctly [element_display_none] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 100 updates animation correctly [element_display_none] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 200 updates animation correctly [element_display_none] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 300 updates animation correctly [element_display_none] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 400 updates animation correctly [element_display_none] assert_equals: expected "0px" but got "100px"
FAIL Scroll at offset 450 updates animation correctly [element_display_none] assert_equals: expected "0px" but got "100px"
Harness: the test ran to completion.
